fix db

+++ src/database/models.py
@@ -1,5 +1,5 @@
 from datetime import datetime, timezone as tz
-from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, LargeBinary, Text
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, ForeignKey, LargeBinary, Text
 from sqlalchemy.orm import relationship
 from src.database.engine import Base
 
@@ -18,7 +18,7 @@ class MessageGroup(Base):
     id = Column(Integer, primary_key=True)
     channel_id = Column(Integer)
     channel_name = Column(String)
-    group_id = Column(Integer)
+    group_id = Column(BigInteger)
     first_message_id = Column(Integer)
     combined_text = Column(Text)
     posted_date = Column(DateTime)
@@ -34,7 +34,7 @@ class Message(Base):
     id = Column(Integer, primary_key=True)
     message_id = Column(Integer)
     text = Column(Text)
-    group_id = Column(Integer, ForeignKey('message_groups.id'))
+    group_id = Column(BigInteger, ForeignKey('message_groups.id'))
     
     group = relationship("MessageGroup", back_populates="messages")
 
@@ -42,7 +42,7 @@ class MediaItem(Base):
     __tablename__ = 'media_items'
 
     id = Column(Integer, primary_key=True)
-    group_id = Column(Integer, ForeignKey('message_groups.id'))
+    group_id = Column(BigInteger, ForeignKey('message_groups.id'))
     media_type = Column(String)  # photo, video, document, etc.
     file_id = Column(String)
     mime_type = Column(String)
